+++ app/code/community/Jtl/Connector/src/Mapper/Product.php
@@ -325,6 +325,9 @@ class Product
 
             if (!is_null($attribute)) {
                 $attribute->setEntityType($product->getResource());
+                if (!$attribute->getSourceModel()) {
+                    $attribute->setSourceModel('eav/entity_attribute_source_table');
+                }
 
                 $values = $attribute
                     ->getSource()
@@ -474,6 +477,18 @@ class Product
 
                 Logger::write('Creating attribute: ' . $attributeCode, Logger::DEBUG);
 
+                // Collect all frontend labels
+                $frontendLabels = array(
+                    \Mage_Core_Model_App::ADMIN_STORE_ID => $attributeName
+                );
+                foreach ($this->stores as $locale => $storeId) {
+                    $variationI18n = ArrayTools::filterOneByLanguage($variation->getI18ns(), LocaleMapper::localeToLanguageIso($locale));
+                    if (!($variationI18n instanceof ConnectorProductVariationI18n))
+                        continue;
+
+                    $frontendLabels[$storeId] = $variationI18n->getName();
+                }
+
                 $attributeData = array(
                     'attribute_code' => $attributeCode,
                     'is_global' => 1,
@@ -494,7 +509,7 @@ class Product
                     'is_required' => 0,
                     'is_visible_in_advanced_search' => 0,
                     'is_visible_on_checkout' => 1,
-                    'frontend_label' => $attributeName,
+                    'frontend_label' => $frontendLabels,
                     'apply_to' => array()
                 );
 
@@ -564,6 +579,9 @@ class Product
 
         $product = \Mage::getModel('catalog/product');
         $attribute->setEntityType($product->getResource());
+        if (!$attribute->getSourceModel()) {
+            $attribute->setSourceModel('eav/entity_attribute_source_table');
+        }
         $values = $attribute
             ->getSource()
             ->getAllOptions(false);
@@ -809,8 +827,13 @@ class Product
 
                 $valueLabels = array();
                 foreach ($stores as $locale => $storeId) {
-                    $valueLabels[$locale] = \Mage::getModel('eav/config')->getAttribute('catalog_product', $attributeOption['attribute_code'])
-                        ->setStoreId($storeId)
+                    $attribute = \Mage::getModel('eav/config')->getAttribute('catalog_product', $attributeOption['attribute_code'])
+                        ->setStoreId($storeId);
+                    if (!$attribute->getSourceModel()) {
+                        $attribute->setSourceModel('eav/entity_attribute_source_table');
+                    }
+
+                    $valueLabels[$locale] = $attribute
                         ->getSource()
                         ->getAllOptions(false);
                 }