+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fix handling of customer's company and VAT number
 - Fix localized attribute names not being set correctly
 - Fix missing simple product name on checkout pages
+- Add VAT ID information to billing addresses
 
 1.1.1.1
 -------